Reorganize design system location
